Added fixed-time Baselines
Fixed time baselines hard-code a time to measure against vs. actually
measuring a baseline.  Use this for determining a strict "operations per
second" on the given hardware.   Note that these benchmarks are
differently comparable to each other vs. ones that make an independent
measurement.

+++ src/Distribution.cpp
@@ -91,10 +91,10 @@ void celero::RunDistribution(uint64_t intArgument)
 					return static_cast<double>(val)/static_cast<double>(delta);
 				}
 
-				return maxVal;
+				return static_cast<double>(maxVal);
 			}
 
-			return minVal;
+			return static_cast<double>(minVal);
 		};
 
 	std::transform(std::begin(series1), std::end(series1), std::begin(series1Normalized),
